requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pygame>=2.5.0
numba>=0.57.0
pillow>=10.0.0
matplotlib
